+++ client.go
@@ -73,8 +73,8 @@ type Client struct {
 	Checkouts        *CheckoutsResource
 	Customers        *CustomersResource
 	Invoices         *InvoicesResource
-	PaymentLinks     *PaymentLinksResource
 	Payments         *PaymentsResource
+	PaymentLinks     *PaymentLinksResource
 	Products         *ProductsResource
 	Subscriptions    *SubscriptionsResource
 	WebhookEndpoints *WebhookEndpointsResource
@@ -129,8 +129,8 @@ func NewClient(apiKey, workspaceID string, opts ...Option) (*Client, error) {
 	c.Checkouts = &CheckoutsResource{http: hc}
 	c.Customers = &CustomersResource{http: hc}
 	c.Invoices = &InvoicesResource{http: hc}
-	c.PaymentLinks = &PaymentLinksResource{http: hc}
 	c.Payments = &PaymentsResource{http: hc}
+	c.PaymentLinks = &PaymentLinksResource{http: hc}
 	c.Products = &ProductsResource{http: hc}
 	c.Subscriptions = &SubscriptionsResource{http: hc}
 	c.WebhookEndpoints = &WebhookEndpointsResource{http: hc}